Time: 0 ms (100%), Space: 7.8 MB (53.3%) - LeetHub

+++ 0231-power-of-two/0231-power-of-two.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPowerOfTwo(int n) {
-        long long i = 1;
+        long i = 1;
         if (n <= 0) {
             return false;
         }
